qdrant-demo/requirements.txt
qdrant-client==1.12.1
sentence-transformers==3.3.1
python-dotenv==1.0.1
